Less pointless casting!

+++ src/main/java/li/cil/tis3d/common/module/ModuleDisplay.java
@@ -182,7 +182,7 @@ public final class ModuleDisplay extends AbstractModuleRotatable {
      *
      * @param value the value that was read.
      */
-    private void process(final int value) {
+    private void process(final short value) {
         drawCall[state.ordinal()] = (byte) value;
         state = state.getNext();
         if (state == State.COLOR) {
